requirements.txt
flask>=3.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
werkzeug>=3.0.0



